requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
python-dotenv>=1.0
google-auth>=2.22
google-auth-oauthlib>=1.1
google-auth-httplib2>=0.1
google-api-python-client>=2.100
requests>=2.31
